# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Varotrabe API",
    "description": "Backend API for Varotrabe E-commerce platform with Supabase integration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "price": { "type": "number" },
          "stock": { "type": "integer" },
          "image_url": { "type": "string" },
          "category": { "type": "string" }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "client_name": { "type": "string" },
          "client_email": { "type": "string", "format": "email" },
          "total_amount": { "type": "number" },
          "status": { "type": "string", "enum": ["pending", "confirmed", "shipped", "delivered", "cancelled"] },
          "created_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
    "/products": {
      "get": {
        "summary": "Get all products",
        "responses": {
          "200": {
            "description": "List of products",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Product" } }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "post": {
        "summary": "Create a new order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "client_name": { "type": "string" },
                  "client_email": { "type": "string" },
                  "total_amount": { "type": "number" },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "product_id": { "type": "string", "format": "uuid" },
                        "quantity": { "type": "integer" },
                        "price": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Order created" }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login success with JWT token" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Request password reset email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "email": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Email sent" }
        }
      }
    },
    "/admin/orders": {
      "get": {
        "summary": "Get all orders (Admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List of all orders" }
        }
      }
    },
    "/admin/orders/{id}/status": {
      "patch": {
        "summary": "Update order status (Admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "type": "object", "properties": { "status": { "type": "string" } } }
            }
          }
        },
        "responses": {
          "200": { "description": "Status updated" }
        }
      }
    },
    "/admin/products": {
      "post": {
        "summary": "Add new product (Admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Product" }
            }
          }
        },
        "responses": {
          "201": { "description": "Product added" }
        }
      }
    },
    "/admin/export/csv": {
      "get": {
        "summary": "Export orders to CSV (Admin)",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "CSV file",
            "content": { "text/csv": {} }
          }
        }
      }
    }
  }
}